create concern module
