+++ config/test.exs
@@ -30,6 +30,8 @@ config :talon, TestTalon.Talon,
   ],
   schema_adapter: Talon.Schema.Adapters.Ecto,
   module: TestTalon,
+  endpoint: Talon.Endpoint,
+  repo: Talon.Repo,
   theme: "admin_lte"
 
 config :talon, TestTalon.FrontEnd,
@@ -38,6 +40,8 @@ config :talon, TestTalon.FrontEnd,
   ],
   schema_adapter: Talon.Schema.Adapters.Ecto,
   module: TestTalon,
+  endpoint: Talon.Endpoint,
+  repo: Talon.Repo,
   theme: "theme2"
 
 config :talon,
