--- test/support/fixtures/talon/admin.ex
@@ -1,4 +1,0 @@
-defmodule TestTalon.Talon do
-  use Talon, otp_app: :talon
-
-end
