+++ test/support/fixtures/talon/talon.ex
@@ -1,4 +1,4 @@
-defmodule TestTalon.FrontEnd do
+defmodule TestTalon.Talon do
   use Talon.Concern, otp_app: :talon
 
 end
